cli | prepare for setuptools
This commit moves the effective cli code to `pycor/cli.py`, into a
function called `main`. This function will eventually be hooked and
run by a shim built by setuptools. For now pycor-cli still exists
as an executable file which simply imports this module.

So the cli is just one way to run our functions - we could build some
Flask routes, run the dispatcher, and we'd have an API.